+++ hotmart_python/__init__.py
@@ -1,1 +1,2 @@
 from .hotmart import Hotmart
+from .decorators import paginate