+++ tests/test_helpers.py
@@ -1,5 +1,3 @@
-import logging
-
 import requests
 import unittest
 from unittest.mock import Mock, MagicMock
@@ -12,12 +10,13 @@ basic = ('Basic YjIzNTQxYzAtMyEzNS20MjVhLWI1ZDItZDM4ZDVkYjcwNGVhOjA5Y2JiMTEz'
          'LWRiZWMtNGI0YS05OWUxLTI3Y2FiNDdkOTI4Mg==')
 
 
-class TestHotmart(unittest.TestCase):
+class TestHelpers(unittest.TestCase):
     def setUp(self):
         self.hotmart = Hotmart(client_id=client_id,
                                client_secret=client_secret,
                                basic=basic)
 
+    # Build Payload
     def test_build_payload_with_valid_arguments(self):
         """
         None values should be ignored
@@ -40,6 +39,22 @@ class TestHotmart(unittest.TestCase):
         result = self.hotmart._build_payload()
         self.assertEqual(result, {})
 
+    # Handle Response
+    def test_handle_response_single_dict_response(self):
+        response = {"key": "value"}
+        expected_result = [response]
+        self.assertEqual(self.hotmart._handle_response(response), expected_result)
+
+    def test_handle_response_list_of_dicts_response(self):
+        response = [{"key1": "value1"}, {"key2": "value2"}]
+        self.assertEqual(self.hotmart._handle_response(response), response)
+
+    def test_handle_response_non_dict_non_list_response(self):
+        response = "invalid_response"
+        with self.assertRaises(ValueError):
+            self.hotmart._handle_response(response) # noqa
+
+    # Build URL
     def test_build_url_with_valid_endpoint_payments(self):
         result = self.hotmart._build_url('payments')
         expected_result = 'https://developers.hotmart.com/payments/api/v1'
@@ -74,6 +89,7 @@ class TestHotmart(unittest.TestCase):
     def test_sandbox_mode_false(self):
         self.assertFalse(self.hotmart.sandbox)
 
+    # Make Request
     @patch('requests.get')
     def test_make_request_successful(self, mock_get):
         mock_response = Mock()
@@ -94,7 +110,7 @@ class TestHotmart(unittest.TestCase):
             }]
         })
 
-    def test_http_error_403_make_request(self):
+    def test_make_request_http_error_403(self):
         """
         Checks if the exception is raised and a log hint it's logged.
         :return:
@@ -120,7 +136,7 @@ class TestHotmart(unittest.TestCase):
         self.assertEqual(e.exception.response.status_code, 403)
         mock_logger.error.assert_called_with('Perhaps the credentials are for Sandbox Mode?')
 
-    def test_http_error_422_make_request(self):
+    def test_make_request_http_error_422(self):
         """
         Checks if the exception is raised and a log hint it's logged.
         :return:
@@ -147,7 +163,7 @@ class TestHotmart(unittest.TestCase):
         mock_logger.error.assert_called_with("This usually happens when the request is missing"
                                              " body parameters.")
 
-    def test_http_error_500_make_request(self):
+    def test_make_request_http_error_500(self):
         """
         Checks if the exception is raised and a log hint it's logged.
         :return:
@@ -175,7 +191,7 @@ class TestHotmart(unittest.TestCase):
         mock_logger.error.assert_any_call('This happens with some endpoints in the Sandbox Mode.')
         mock_logger.error.assert_any_call('Usually the API it\'s not down, it\'s just a bug.')
 
-    def test_http_general_error_make_request(self):
+    def test_make_request_http_general_error(self):
         """
         Checks if the exception is raised when errors are different from 401, 403, 422 or 500.
         :return:
@@ -200,20 +216,22 @@ class TestHotmart(unittest.TestCase):
         self.assertEqual(e.exception.response.status_code, 418)
         self.assertRaises(requests.exceptions.HTTPError)
 
+    # Is token expired
     @patch('time.time')
-    def test_token_expired_when_expiry_in_past(self, mock_time):
+    def test_is_token_expired_when_expiry_in_past(self, mock_time):
         mock_time.return_value = 100
         self.hotmart.token_expires_at = 99
         self.assertTrue(self.hotmart._is_token_expired())
 
     @patch('time.time')
-    def test_token_not_expired_when_expiry_in_future(self, mock_time):
+    def test_is_token_expired_when_expiry_in_future(self, mock_time):
         mock_time.return_value = 100
         self.hotmart.token_expires_at = 101
         self.assertFalse(self.hotmart._is_token_expired())
 
+    # Fetch new token
     @patch.object(Hotmart, '_make_request')
-    def test_token_obtained_successfully(self, mock_make_request):
+    def test_fetch_new_token_obtained_successfully(self, mock_make_request):
         mock_make_request.return_value = {
             'access_token': 'test_token'
         }
@@ -221,14 +239,15 @@ class TestHotmart(unittest.TestCase):
         self.assertEqual(token, 'test_token')
 
     @patch('requests.post')
-    def test_token_obtained_failure(self, mock_post):
+    def test_fetch_new_token_obtained_failure(self, mock_post):
         mock_post.side_effect = requests.exceptions.RequestException
         with self.assertRaises(requests.exceptions.RequestException):
             self.hotmart._fetch_new_token()
 
+    # Get token
     @patch.object(Hotmart, '_is_token_expired')
     @patch.object(Hotmart, '_fetch_new_token')
-    def test_token_found_in_cache(self, mock_fetch_new_token, mock_is_token_expired):
+    def test_get_token_found_in_cache(self, mock_fetch_new_token, mock_is_token_expired):
 
         mock_is_token_expired.return_value = False
         self.hotmart.token_cache = 'test_token'
@@ -239,9 +258,9 @@ class TestHotmart(unittest.TestCase):
 
     @patch.object(Hotmart, '_is_token_expired')
     @patch.object(Hotmart, '_fetch_new_token')
-    def test_token_not_in_cache_and_fetched_success(self,
-                                                    mock_fetch_new_token,
-                                                    mock_is_token_expired):
+    def test_get_token_not_in_cache_and_fetched_success(self,
+                                                        mock_fetch_new_token,
+                                                        mock_is_token_expired):
 
         mock_is_token_expired.return_value = True
         mock_fetch_new_token.return_value = 'new_token'
@@ -250,25 +269,28 @@ class TestHotmart(unittest.TestCase):
 
     @patch.object(Hotmart, '_is_token_expired')
     @patch.object(Hotmart, '_fetch_new_token')
-    def test_token_not_in_cache_and_fetch_failed(self, mock_fetch_new_token, mock_is_token_expired):
+    def test_get_token_not_in_cache_and_fetch_failed(self,
+                                                     mock_fetch_new_token,
+                                                     mock_is_token_expired):
 
         mock_is_token_expired.return_value = True
         mock_fetch_new_token.return_value = None
         token = self.hotmart._get_token()
         self.assertIsNone(token)
 
+    # Request with token
     @patch.object(Hotmart, '_get_token')
     @patch.object(Hotmart, '_make_request')
-    def test_successful_request_with_token(self, mock_make_request, mock_get_token):
+    def test_request_with_token_successful(self, mock_make_request, mock_get_token):
 
         mock_get_token.return_value = 'test_token'
         mock_make_request.return_value = {"success": True}
         result = self.hotmart._request_with_token('GET', 'https://example.com')
-        self.assertEqual(result, {"success": True})
+        self.assertEqual(result, [{"success": True}])
 
     @patch.object(Hotmart, '_get_token')
     @patch.object(Hotmart, '_make_request')
-    def test_failed_request_with_token(self, mock_make_request, mock_get_token):
+    def test_request_with_token_failed(self, mock_make_request, mock_get_token):
 
         mock_get_token.return_value = 'test_token'
         mock_make_request.side_effect = requests.exceptions.RequestException
@@ -276,66 +298,11 @@ class TestHotmart(unittest.TestCase):
             self.hotmart._request_with_token('GET', 'https://example.com')
 
     @patch.object(Hotmart, '_get_token')
-    def test_unsupported_method_with_token(self, mock_get_token):
+    def test_request_with_token_unsupported_method(self, mock_get_token):
 
         mock_get_token.return_value = 'test_token'
         with self.assertRaises(ValueError):
             self.hotmart._request_with_token('PUT', 'https://example.com')
-
-    @patch.object(Hotmart, '_request_with_token')
-    def test_pagination_without_pagination(self, mock_request_with_token):
-
-        mock_request_with_token.return_value = {
-            "items": ["item1", "item2"]
-        }
-        result = self.hotmart._pagination('GET',
-                                          'https://example.com')
-
-        self.assertEqual(result, ["item1", "item2"])
-
-    @patch.object(Hotmart, '_request_with_token')
-    def test_pagination_with_single_page(self, mock_request_with_token):
-
-        mock_request_with_token.return_value = {
-            "items": ["item1", "item2"],
-            "page_info": {}
-        }
-
-        result = self.hotmart._pagination('GET',
-                                          'https://example.com',
-                                          paginate=True)
-
-        self.assertEqual(result, ["item1", "item2"])
-
-    @patch.object(Hotmart, '_request_with_token')
-    def test_pagination_with_multiple_pages(self, mock_request_with_token):
-
-        mock_request_with_token.side_effect = [
-            {
-                "items": ["item1", "item2"],
-                "page_info": {"next_page_token": "token"}
-            },
-            {
-                "items": ["item3", "item4"],
-                "page_info": {}
-            }
-        ]
-        params = {}
-        result = self.hotmart._pagination('GET',
-                                          'https://example.com',
-                                          params=params,
-                                          paginate=True)
-
-        self.assertEqual(result, ["item1", "item2", "item3", "item4"])
-
-    @patch.object(Hotmart, '_request_with_token')
-    def test_pagination_with_failed_first_page(self, mock_request_with_token):
-        mock_request_with_token.return_value = None
-
-        with self.assertRaises(ValueError):
-            self.hotmart._pagination('GET',
-                                     'https://example.com',
-                                     paginate=True)
 
 
 if __name__ == '__main__':